+++ paddle/phi/core/distributed/store/CMakeLists.txt
@@ -3,6 +3,13 @@ cc_library(
   SRCS tcp_store.cc tcp_utils.cc socket.cpp store.cc
   DEPS enforce glog)
 
+if(WITH_GLOO)
+  cc_library(
+    gloo_store
+    SRCS gloo_store.cc
+    DEPS gloo)
+endif()
+
 if(NOT WIN32)
   cc_test(
     test_c_tcp_store
